* Afegits atributs identificadors d'event i resource als bookings.  * Arreglats els tests que fallaven de la classe Event.  * Actualitzats tests de SimpleBooking.  * Implementant RecurrentBooking i els seus tests.
git-svn-id: http://devel.cpl.upc.edu/recursos/svn/smeagol/java/branches/t251@1318 4ab63a03-c2d3-4037-8559-9a21b594da05

+++ smeagol-client/src/test/java/edu/upc/cpl/smeagol/client/domain/ResourceTest.java
@@ -6,8 +6,8 @@ import java.util.Collection;
 import junit.framework.TestCase;
 
 public class ResourceTest extends TestCase {
-	private static final int ID1 = 1;
-	private static final int ID2 = 2;
+	private static final long ID1 = 1;
+	private static final long ID2 = 2;
 	private static final String DESC1 = "description1";
 	private static final String DESC2 = "description2";
 	private static final String INFO1 = "chachi1";
